Work on optimizing the frame rate. Needs testing.
This is done in order to enable multi frame capture to reduce
noise. The improvement is by increasing the pixelrate.
Note that it the camera pixelrate is limited to 25(Mpix) also by
the use of the USB to connect to the Arduino. Without the sunshader
the pixelrate can reach above 30.